try o1 suggestion
