--- src/__tests__/__mocks__/PolicyEngine.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const MockPolicyEngine = () => {
-  return <div data-testid="mock-policy-engine">Mocked PolicyEngine</div>;
-};
-
-export default MockPolicyEngine;
